requirements.txt
boto3
awswrangler
python-dotenv
pandas